requirements.txt
requests
krakenex
python-dotenv